+++ bank-feeds/RELEASES.md
@@ -139,3 +139,13 @@ Based on:
 - [csharp v2.3.0] bank-feeds
 ### Releases
 - [NuGet v2.3.0] https://www.nuget.org/packages/Codat.BankFeeds/2.3.0 - bank-feeds
+
+## 2023-10-02 00:13:02
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.92.3 (2.143.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v2.3.1] bank-feeds
+### Releases
+- [NuGet v2.3.1] https://www.nuget.org/packages/Codat.BankFeeds/2.3.1 - bank-feeds